Abolish provers/tests/provers as this is the only module in this crate.
