+++ provers/perf/src/lib.rs
@@ -1,5 +1,5 @@
 use strata_zkvm::{ProofType, ZkVmHost, ZkVmInputBuilder, ZkVmProver, ZkVmResult};
-use strata_zkvm_tests::provers::{
+use strata_zkvm_tests::{
     btc::BtcBlockProofGenerator, checkpoint::CheckpointProofGenerator, cl::ClProofGenerator,
     el::ElProofGenerator, l1_batch::L1BatchProofGenerator, l2_batch::L2BatchProofGenerator,
     ProofGenerator,
