+++ provers/tests/src/btc.rs
@@ -56,21 +56,18 @@ mod test {
     #[test]
     #[cfg(not(any(feature = "risc0", feature = "sp1")))]
     fn test_native() {
-        use crate::provers::TEST_NATIVE_GENERATORS;
-        test_proof(TEST_NATIVE_GENERATORS.btc_blockspace());
+        test_proof(crate::provers::TEST_NATIVE_GENERATORS.btc_blockspace());
     }
 
     #[test]
     #[cfg(feature = "risc0")]
     fn test_risc0() {
-        use crate::provers::TEST_RISC0_GENERATORS;
-        test_proof(TEST_RISC0_GENERATORS.btc_blockspace());
+        test_proof(crate::TEST_RISC0_GENERATORS.btc_blockspace());
     }
 
     #[test]
     #[cfg(feature = "sp1")]
     fn test_sp1() {
-        use crate::provers::TEST_SP1_GENERATORS;
-        test_proof(TEST_SP1_GENERATORS.btc_blockspace());
+        test_proof(crate::TEST_SP1_GENERATORS.btc_blockspace());
     }
 }
