+++ provers/tests/src/checkpoint.rs
@@ -107,21 +107,18 @@ mod test {
 
     #[test]
     fn test_native() {
-        use crate::provers::TEST_NATIVE_GENERATORS;
-        test_proof(TEST_NATIVE_GENERATORS.checkpoint());
+        test_proof(crate::TEST_NATIVE_GENERATORS.checkpoint());
     }
 
     #[test]
     #[cfg(feature = "risc0")]
     fn test_risc0() {
-        use crate::provers::TEST_RISC0_GENERATORS;
-        test_proof(TEST_RISC0_GENERATORS.checkpoint());
+        test_proof(crate::TEST_RISC0_GENERATORS.checkpoint());
     }
 
     #[test]
     #[cfg(feature = "sp1")]
     fn test_sp1() {
-        use crate::provers::TEST_SP1_GENERATORS;
-        test_proof(TEST_SP1_GENERATORS.checkpoint());
+        test_proof(crate::TEST_SP1_GENERATORS.checkpoint());
     }
 }
